优化缓存元素的过期时间处理逻辑, 完善对@ResponseCache注解处理过程, 提供IWebCacheProcessor接口的默认实现以支持对视图执行结果缓存处理, 另外将缓存超时时间配置的单位又调整为秒了:p

+++ ymate-platform-cache/src/main/java/net/ymate/platform/cache/ICacheModuleCfg.java
@@ -54,7 +54,7 @@ public interface ICacheModuleCfg {
     String getDefaultCacheName();
 
     /**
-     * @return 缓存数据超时时间，可选参数，数值必须大于等于0，为0表示默认缓存300000毫秒
+     * @return 缓存数据超时时间，可选参数，数值必须大于等于0，为0表示默认缓存300毫秒
      */
-    long getDefaultCacheTimeout();
+    int getDefaultCacheTimeout();
 }
